third_party/blink/web_tests/fast/dom/MutationObserver/mutation-record-nullity-expected.txt

Non-relevant properties on mutation records should be null, except for NodeLists, which should be empty

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

characterData record:
PASS record.attributeName is null
PASS record.attributeNamespace is null
PASS record.oldValue is null
PASS record.previousSibling is null
PASS record.nextSibling is null
PASS record.addedNodes.length is 0
PASS record.removedNodes.length is 0

childList record:
PASS record.attributeName is null
PASS record.attributeNamespace is null
PASS record.oldValue is null

attributes record:
PASS record.attributeNamespace is null
PASS record.oldValue is null
PASS record.previousSibling is null
PASS record.nextSibling is null
PASS record.addedNodes.length is 0
PASS record.removedNodes.length is 0
PASS successfullyParsed is true

TEST COMPLETE
